Update to use "deprecated" version 3.0.0 to get rid of "already initialized constant Deprecate" warning.
rubygems defines its own Deprecate module in
rubygems/deprecate.rb. This causes a warning when deprecate 2.0.1 is
required because it aliases the Deprecated module to
Deprecate. Version 3.0.0 no longer does this aliasing and gets rid of
the warning. Updating to 3.0.0 requires some changes in how methods
are deprecated, and how set_action is called.